+++ packages/theme/package.json
@@ -10,6 +10,9 @@
     ".": {
       "import": "./dist/index.js"
     },
+    "./provider": {
+      "import": "./dist/components/ThemeProvider.js"
+    },
     "./styles": "./dist/index.css"
   },
   "scripts": {
