+++ packages/theme/src/index.ts
@@ -1,4 +1,20 @@
-export * from './themes';
-export * from './sprinkles';
-export * from './tokens';
-export * from './components/ThemeProvider';
+export { vars } from './themes/themes.css';
+export type { VarsType } from './themes/themes.css';
+export type { ThemeContract } from './themes/contract';
+
+export { sprinkles } from './sprinkles/sprinkles.css';
+export type { Sprinkles } from './sprinkles/sprinkles.css';
+
+export { ThemeProvider } from './components/ThemeProvider';
+
+export { tokens } from './tokens/tokens';
+export type { TokensType } from './tokens/tokens';
+
+export { colors } from './tokens/colors';
+export type { ColorsType } from './tokens/colors';
+export { radius } from './tokens/radius';
+export type { RadiusType } from './tokens/radius';
+export { spacing } from './tokens/spacing';
+export type { SpacingType } from './tokens/spacing';
+export { typography } from './tokens/typography';
+export type { TypographyType } from './tokens/typography';
