--- packages/theme/src/sprinkles/index.ts
@@ -1,1 +1,0 @@
-export { sprinkles } from './sprinkles.css';
