+++ packages/theme/src/themes/contract.ts
@@ -1,4 +1,4 @@
-import { tokens } from '../tokens';
+import { tokens } from '@/tokens/tokens';
 
 export type ThemeContract = {
   colors: {
