+++ packages/theme/src/themes/dark.ts
@@ -1,4 +1,4 @@
-import { tokens } from '../tokens';
+import { tokens } from '../tokens/tokens';
 import type { ThemeContract } from './contract';
 
 export const darkTheme: ThemeContract = {
