--- packages/theme/src/themes/index.ts
@@ -1,5 +1,0 @@
-import { vars } from './themes.css';
-
-export * from './themes.css';
-export type { ThemeContract } from './contract';
-export type VarsType = typeof vars;
