+++ packages/theme/src/themes/themes.css.ts
@@ -5,3 +5,4 @@ import type { ThemeContract } from './contract';
 
 export const [lightThemeClass, vars] = createTheme<ThemeContract>(lightTheme);
 export const darkThemeClass = createTheme(vars, darkTheme);
+export type VarsType = typeof vars;
