+++ packages/theme/src/tokens/typography.ts
@@ -14,3 +14,5 @@ export const typography = {
     bold: '700',
   },
 } as const;
+
+export type TypographyType = typeof typography;
